# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Buggy Insurance backend API",
        "title": "Buggy Insurance Backend API",
        "contact": {
            "name": "Alexander Rostovshchikov",
            "email": "rostovshchikov.fortech@gmail.com"
        },
        "version": "1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/applications": {
            "get": {
                "description": "Retrieve a paginated list of applications for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Applications"
                ],
                "summary": "Get applications list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {accessToken}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.GetApplicationsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request — invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized — user not logged in",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found — resource not found",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity — validation error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new application for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Applications"
                ],
                "summary": "Create a new application",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {accessToken}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Application data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateApplicationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Application"
                        }
                    },
                    "400": {
                        "description": "Bad Request — invalid request body",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized — user not logged in",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found — resource not found",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict — resource conflict (e.g. duplicate)",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity — validation error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/manager/applications/{id}/comments": {
            "post": {
                "description": "Allows a manager to add a comment to a specific application",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manager"
                ],
                "summary": "Create a comment for an application",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {accessToken}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Application ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Comment payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateApplicationCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.CreateApplicationCommentResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid application id or request body",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Manager or application not found",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Comment is required",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/manager/dashboard": {
            "get": {
                "description": "Retrieve overview stats for the manager",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manager"
                ],
                "summary": "Get manager dashboard",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {accessToken}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.ManagerDashboardResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve dashboard",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/manager/statistics": {
            "get": {
                "description": "Retrieve application statistics by period (week, month, quarter, year)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manager"
                ],
                "summary": "Get manager statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {accessToken}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "month",
                        "description": "Period filter: week, month, quarter, year",
                        "name": "period",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.ManagerStatisticsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Invalid period",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve statistics",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Application": {
            "type": "object",
            "properties": {
                "calculatedPrice": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "productType": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "domain.ChartDataEntry": {
            "type": "object",
            "properties": {
                "approved": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "new": {
                    "type": "integer"
                },
                "rejected": {
                    "type": "integer"
                }
            }
        },
...
        "domain.CommentAuthor": {
            "type": "object",
            "properties": {
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "domain.ConversionStats": {
            "type": "object",
            "properties": {
                "approved": {
                    "type": "integer"
                },
                "rate": {
                    "type": "number"
                },
                "rejected": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "domain.CreateApplicationCommentRequest": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                }
            }
        },
        "domain.CreateApplicationCommentResponse": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/domain.CommentAuthor"
                },
                "comment": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "domain.CreateApplicationRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "managerId": {
                    "type": "integer"
                },
                "productId": {
                    "type": "integer"
                },
                "productType": {
                    "type": "string"
                }
            }
        },
...
        "domain.Details": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "$ref": "#/definitions/domain.Details"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "domain.GetApplicationsResponse": {
            "type": "object",
            "properties": {
                "applications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Application"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/domain.Pagination"
                }
            }
        },
...
        "domain.ManagerDashboardResponse": {
            "type": "object",
            "properties": {
                "chartData": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ChartDataEntry"
                    }
                },
                "recentApplications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.RecentApplicationEntry"
                    }
                },
                "stats": {
                    "$ref": "#/definitions/domain.ManagerDashboardStats"
                }
            }
        },
        "domain.ManagerDashboardStats": {
            "type": "object",
            "properties": {
                "approvedThisMonth": {
                    "type": "integer"
                },
                "newToday": {
                    "type": "integer"
                },
                "rejectedThisMonth": {
                    "type": "integer"
                },
                "underReview": {
                    "type": "integer"
                }
            }
        },
        "domain.ManagerStatisticsResponse": {
            "type": "object",
            "properties": {
                "byStatus": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "byType": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "conversion": {
                    "$ref": "#/definitions/domain.ConversionStats"
                },
                "period": {
                    "type": "string"
                }
            }
        },
        "domain.Pagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "domain.RecentApplicationEntry": {
            "type": "object",
            "properties": {
                "calculatedPrice": {
                    "type": "integer"
                },
                "clientFullName": {
                    "type": "string"
                },
                "clientId": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "productType": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}